+++ src/loot_of_lima_solver/board.py
