+++ src/loot_of_lima_solver/enums.py
