+++ src/loot_of_lima_solver/exceptions.py
