+++ src/loot_of_lima_solver/game.py
