+++ src/loot_of_lima_solver/information.py
